added my trips page
Added my trips page to display resturant, hotel, and flight information based on a given location. I was able to get the Google Places API working and set up the search autocomplete along with the dropdowns for hotels and airlines. I want to make it so that the user can save the resturant to their trip on the column on the right by clicking the plus sign on the resturant. We can send this data to a database to save for the common deliverable.